chore: move test location
